requirements.txt
Flask>2.3
Werkzeug>2.3
PyYAML>6.0
